vscroll fix
git-svn-id: https://svn.code.sf.net/p/sauerbraten/code/src@2298 07af1644-1164-4067-b897-63bf4dc9b748

+++ engine/octaedit.cpp
@@ -1506,7 +1506,7 @@ void voffset(int *x, int *y)
 }
 COMMAND(voffset, "ii");
 
-void vscroll(float *s, int *t)
+void vscroll(float *s, float *t)
 {
     if(noedit() || (nompedit && multiplayer())) return;
     VSlot ds;
